requirements.txt
ctns>=0.4.2
networkx>=2.4
dash_gif_component>=1.0.2
dash>=1.11.0
dash_html_components>=1.0.3
dash_bootstrap_components>=0.9.2
dash_core_components>=1.9.1
dash_cytoscape>=0.0.5
gunicorn>=19.9.0
matplotlib>=3.0.3
Pillow>=7.1.2
plotly>=4.7.1
python_igraph>=0.8.2
requests>=2.23.0
tqdm>=4.46.1